Decoupled the waiting room logic from the exam.py router
